requirements/dev.txt
asgiref==3.8.1
attrs==24.3.0
Django==5.1
django-autoslug==1.9.9
django-cities-light==3.10.1
django-cors-headers==4.6.0
django-phonenumber-field==8.0.0
django-tinymce==4.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
pillow==11.0.0
progressbar2==4.5.0
PyJWT==2.10.1
python-utils==3.9.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.36.1
rpds-py==0.22.3
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
Unidecode==1.3.8
uritemplate==4.1.1
whitenoise==6.8.2
